third_party/WebKit/LayoutTests/html/text_level_semantics/a-tooltip-null.html

<!DOCTYPE html>
<html>
<body onload="test()">
<script src="../../resources/js-test.js"></script>
<span id="target" href="#" title="This should not appear as a tooltip"><span title>Hello.</span></span>
<script>
window.jsTestIsAsync = true;

function test() {
    if (window.eventSender) {
        eventSender.dragMode = false;
        eventSender.mouseMoveTo(target.offsetLeft, target.offsetTop + target.offsetHeight / 2);
        shouldBeEqualToString("testRunner.tooltipText", "");

        target.style.display = "none";
        finishJSTest();
    }
}
</script>
</body>
</html>
