third_party/blink/web_tests/html/text_level_semantics/a-tooltip-null-expected.txt

PASS testRunner.tooltipText is ""
PASS successfullyParsed is true

TEST COMPLETE
